Select class is created for separating filtering and concatenation logic from Concat.
